Move code to pages.rs
